Rename X11_resources -> x11_resources
