+++ src/platforms/x11/input/input.cpp
@@ -17,7 +17,7 @@
  */
 
 #include "input_platform.h"
-#include "../X11_resources.h"
+#include "../x11_resources.h"
 #include "mir/module_properties.h"
 #include "mir/assert_module_entry_point.h"
 #include "mir/libname.h"
