+++ src/platforms/x11/x11_resources.h
